refactor(store): extract shared types, StoreInterface, and format package
Prepare the codebase for cloud sync by extracting shared abstractions:

- Create internal/types with all shared data types (Observation, Session,
  SearchResult, Stats, etc.) — zero dependencies on internal packages
- Define StoreInterface (StoreReader + StoreWriter) and StoreSyncer in
  internal/types — both *store.Store and future RemoteStore implement these
- Add compile-time interface satisfaction checks for *Store
- Use type aliases in store package for full backward compatibility —
  all existing store.Observation references continue to work unchanged
- Extract FormatContext logic to internal/format shared package —
  Store.FormatContext now delegates to format.Context()
- Add 6 tests for the format package (empty, sessions, observations,
  prompts, full output, truncation)
