+++ internal/store/store.go
@@ -21,143 +21,47 @@ import (
 	"time"
 
 	sqlite "modernc.org/sqlite"
+
+	"github.com/Gentleman-Programming/engram/internal/format"
+	"github.com/Gentleman-Programming/engram/internal/types"
 )
 
 var openDB = sql.Open
+
+// Compile-time check: *Store must satisfy types.StoreInterface and types.StoreSyncer.
+var (
+	_ types.StoreInterface = (*Store)(nil)
+	_ types.StoreSyncer    = (*Store)(nil)
+)
 
 // sqliteConstraintForeignKey is the extended SQLite result code for a foreign-key
 // constraint violation (SQLITE_CONSTRAINT_FOREIGNKEY = 787).
 // See https://www.sqlite.org/rescode.html#constraint_foreignkey
 const sqliteConstraintForeignKey = 787
 
-// Sentinel errors returned by delete operations so callers can use errors.Is.
+// Sentinel errors — aliases to types package for backward compatibility.
 var (
-	ErrSessionNotFound        = errors.New("session not found")
-	ErrSessionHasObservations = errors.New("session still has observations")
-	ErrPromptNotFound         = errors.New("prompt not found")
+	ErrSessionNotFound        = types.ErrSessionNotFound
+	ErrSessionHasObservations = types.ErrSessionHasObservations
+	ErrPromptNotFound         = types.ErrPromptNotFound
 )
 
-// ─── Types ───────────────────────────────────────────────────────────────────
+// ─── Types (aliases to internal/types for backward compatibility) ────────────
+// All type definitions live in internal/types. These aliases ensure that
+// existing code using store.Observation, store.Session, etc. continues to work.
 
-type Session struct {
-	ID        string  `json:"id"`
-	Project   string  `json:"project"`
-	Directory string  `json:"directory"`
-	StartedAt string  `json:"started_at"`
-	EndedAt   *string `json:"ended_at,omitempty"`
-	Summary   *string `json:"summary,omitempty"`
-}
-
-type Observation struct {
-	ID             int64   `json:"id"`
-	SyncID         string  `json:"sync_id"`
-	SessionID      string  `json:"session_id"`
-	Type           string  `json:"type"`
-	Title          string  `json:"title"`
-	Content        string  `json:"content"`
-	ToolName       *string `json:"tool_name,omitempty"`
-	Project        *string `json:"project,omitempty"`
-	Scope          string  `json:"scope"`
-	TopicKey       *string `json:"topic_key,omitempty"`
-	CreatedBy      string  `json:"created_by,omitempty"`
-	UpdatedBy      string  `json:"updated_by,omitempty"`
-	RevisionCount  int     `json:"revision_count"`
-	DuplicateCount int     `json:"duplicate_count"`
-	LastSeenAt     *string `json:"last_seen_at,omitempty"`
-	CreatedAt      string  `json:"created_at"`
-	UpdatedAt      string  `json:"updated_at"`
-	DeletedAt      *string `json:"deleted_at,omitempty"`
-}
-
-type SearchResult struct {
-	Observation
-	Rank float64 `json:"rank"`
-}
-
-type SessionSummary struct {
-	ID               string  `json:"id"`
-	Project          string  `json:"project"`
-	StartedAt        string  `json:"started_at"`
-	EndedAt          *string `json:"ended_at,omitempty"`
-	Summary          *string `json:"summary,omitempty"`
-	ObservationCount int     `json:"observation_count"`
-}
-
-type Stats struct {
-	TotalSessions     int      `json:"total_sessions"`
-	TotalObservations int      `json:"total_observations"`
-	TotalPrompts      int      `json:"total_prompts"`
-	Projects          []string `json:"projects"`
-}
-
-type TimelineEntry struct {
-	ID             int64   `json:"id"`
-	SessionID      string  `json:"session_id"`
-	Type           string  `json:"type"`
-	Title          string  `json:"title"`
-	Content        string  `json:"content"`
-	ToolName       *string `json:"tool_name,omitempty"`
-	Project        *string `json:"project,omitempty"`
-	Scope          string  `json:"scope"`
-	TopicKey       *string `json:"topic_key,omitempty"`
-	RevisionCount  int     `json:"revision_count"`
-	DuplicateCount int     `json:"duplicate_count"`
-	LastSeenAt     *string `json:"last_seen_at,omitempty"`
-	CreatedAt      string  `json:"created_at"`
-	UpdatedAt      string  `json:"updated_at"`
-	DeletedAt      *string `json:"deleted_at,omitempty"`
-	IsFocus        bool    `json:"is_focus"` // true for the anchor observation
-}
-
-type TimelineResult struct {
-	Focus        Observation     `json:"focus"`        // The anchor observation
-	Before       []TimelineEntry `json:"before"`       // Observations before the focus (chronological)
-	After        []TimelineEntry `json:"after"`        // Observations after the focus (chronological)
-	SessionInfo  *Session        `json:"session_info"` // Session that contains the focus observation
-	TotalInRange int             `json:"total_in_range"`
-}
-
-type SearchOptions struct {
-	Type    string `json:"type,omitempty"`
-	Project string `json:"project,omitempty"`
-	Scope   string `json:"scope,omitempty"`
-	Limit   int    `json:"limit,omitempty"`
-}
-
-type AddObservationParams struct {
-	SessionID string `json:"session_id"`
-	Type      string `json:"type"`
-	Title     string `json:"title"`
-	Content   string `json:"content"`
-	ToolName  string `json:"tool_name,omitempty"`
-	Project   string `json:"project,omitempty"`
-	Scope     string `json:"scope,omitempty"`
-	TopicKey  string `json:"topic_key,omitempty"`
-}
-
-type UpdateObservationParams struct {
-	Type     *string `json:"type,omitempty"`
-	Title    *string `json:"title,omitempty"`
-	Content  *string `json:"content,omitempty"`
-	Project  *string `json:"project,omitempty"`
-	Scope    *string `json:"scope,omitempty"`
-	TopicKey *string `json:"topic_key,omitempty"`
-}
-
-type Prompt struct {
-	ID        int64  `json:"id"`
-	SyncID    string `json:"sync_id"`
-	SessionID string `json:"session_id"`
-	Content   string `json:"content"`
-	Project   string `json:"project,omitempty"`
-	CreatedAt string `json:"created_at"`
-}
-
-type AddPromptParams struct {
-	SessionID string `json:"session_id"`
-	Content   string `json:"content"`
-	Project   string `json:"project,omitempty"`
-}
+type Session = types.Session
+type Observation = types.Observation
+type SearchResult = types.SearchResult
+type SessionSummary = types.SessionSummary
+type Stats = types.Stats
+type TimelineEntry = types.TimelineEntry
+type TimelineResult = types.TimelineResult
+type SearchOptions = types.SearchOptions
+type AddObservationParams = types.AddObservationParams
+type UpdateObservationParams = types.UpdateObservationParams
+type Prompt = types.Prompt
+type AddPromptParams = types.AddPromptParams
 
 const (
 	DefaultSyncTargetKey = "cloud"
@@ -179,38 +83,9 @@ const (
 	SyncSourceRemote = "remote"
 )
 
-type SyncState struct {
-	TargetKey           string  `json:"target_key"`
-	Lifecycle           string  `json:"lifecycle"`
-	LastEnqueuedSeq     int64   `json:"last_enqueued_seq"`
-	LastAckedSeq        int64   `json:"last_acked_seq"`
-	LastPulledSeq       int64   `json:"last_pulled_seq"`
-	ConsecutiveFailures int     `json:"consecutive_failures"`
-	BackoffUntil        *string `json:"backoff_until,omitempty"`
-	LeaseOwner          *string `json:"lease_owner,omitempty"`
-	LeaseUntil          *string `json:"lease_until,omitempty"`
-	LastError           *string `json:"last_error,omitempty"`
-	UpdatedAt           string  `json:"updated_at"`
-}
-
-type SyncMutation struct {
-	Seq        int64   `json:"seq"`
-	TargetKey  string  `json:"target_key"`
-	Entity     string  `json:"entity"`
-	EntityKey  string  `json:"entity_key"`
-	Op         string  `json:"op"`
-	Payload    string  `json:"payload"`
-	Source     string  `json:"source"`
-	Project    string  `json:"project"`
-	OccurredAt string  `json:"occurred_at"`
-	AckedAt    *string `json:"acked_at,omitempty"`
-}
-
-// EnrolledProject represents a project enrolled for cloud sync.
-type EnrolledProject struct {
-	Project    string `json:"project"`
-	EnrolledAt string `json:"enrolled_at"`
-}
+type SyncState = types.SyncState
+type SyncMutation = types.SyncMutation
+type EnrolledProject = types.EnrolledProject
 
 type syncSessionPayload struct {
 	ID        string  `json:"id"`
@@ -245,13 +120,7 @@ type syncPromptPayload struct {
 }
 
 // ExportData is the full serializable dump of the engram database.
-type ExportData struct {
-	Version      string        `json:"version"`
-	ExportedAt   string        `json:"exported_at"`
-	Sessions     []Session     `json:"sessions"`
-	Observations []Observation `json:"observations"`
-	Prompts      []Prompt      `json:"prompts"`
-}
+type ExportData = types.ExportData
 
 // ─── Config ──────────────────────────────────────────────────────────────────
 
@@ -1729,44 +1598,7 @@ func (s *Store) FormatContext(project, scope string) (string, error) {
 		return "", err
 	}
 
-	if len(sessions) == 0 && len(observations) == 0 && len(prompts) == 0 {
-		return "", nil
-	}
-
-	var b strings.Builder
-	b.WriteString("## Memory from Previous Sessions\n\n")
-
-	if len(sessions) > 0 {
-		b.WriteString("### Recent Sessions\n")
-		for _, sess := range sessions {
-			summary := ""
-			if sess.Summary != nil {
-				summary = fmt.Sprintf(": %s", truncate(*sess.Summary, 200))
-			}
-			fmt.Fprintf(&b, "- **%s** (%s)%s [%d observations]\n",
-				sess.Project, sess.StartedAt, summary, sess.ObservationCount)
-		}
-		b.WriteString("\n")
-	}
-
-	if len(prompts) > 0 {
-		b.WriteString("### Recent User Prompts\n")
-		for _, p := range prompts {
-			fmt.Fprintf(&b, "- %s: %s\n", p.CreatedAt, truncate(p.Content, 200))
-		}
-		b.WriteString("\n")
-	}
-
-	if len(observations) > 0 {
-		b.WriteString("### Recent Observations\n")
-		for _, obs := range observations {
-			fmt.Fprintf(&b, "- [%s] **%s**: %s\n",
-				obs.Type, obs.Title, truncate(obs.Content, 300))
-		}
-		b.WriteString("\n")
-	}
-
-	return b.String(), nil
+	return format.Context(sessions, observations, prompts), nil
 }
 
 // ─── Export / Import ─────────────────────────────────────────────────────────
@@ -1945,12 +1777,7 @@ func (s *Store) Import(data *ExportData) (*ImportResult, error) {
 	return result, nil
 }
 
-type ImportResult struct {
-	SessionsImported     int `json:"sessions_imported"`
-	ObservationsImported int `json:"observations_imported"`
-	ObservationsUpdated  int `json:"observations_updated"`
-	PromptsImported      int `json:"prompts_imported"`
-}
+type ImportResult = types.ImportResult
 
 // ─── Sync Chunk Tracking ─────────────────────────────────────────────────────
 
@@ -2346,12 +2173,7 @@ func (s *Store) IsProjectEnrolled(project string) (bool, error) {
 
 // ─── Project Migration ───────────────────────────────────────────────────────
 
-type MigrateResult struct {
-	Migrated            bool  `json:"migrated"`
-	ObservationsUpdated int64 `json:"observations_updated"`
-	SessionsUpdated     int64 `json:"sessions_updated"`
-	PromptsUpdated      int64 `json:"prompts_updated"`
-}
+type MigrateResult = types.MigrateResult
 
 func (s *Store) MigrateProject(oldName, newName string) (*MigrateResult, error) {
 	if oldName == "" || newName == "" || oldName == newName {
@@ -2442,13 +2264,7 @@ func (s *Store) ListProjectNames() ([]string, error) {
 }
 
 // ProjectStats holds aggregate statistics for a single project.
-type ProjectStats struct {
-	Name             string   `json:"name"`
-	ObservationCount int      `json:"observation_count"`
-	SessionCount     int      `json:"session_count"`
-	PromptCount      int      `json:"prompt_count"`
-	Directories      []string `json:"directories"` // unique directories from sessions
-}
+type ProjectStats = types.ProjectStats
 
 // ListProjectsWithStats returns all projects with aggregated counts.
 // Ordered by observation count descending.
@@ -3620,19 +3436,10 @@ func sanitizeFTS(query string) string {
 // ─── Passive Capture ─────────────────────────────────────────────────────────
 
 // PassiveCaptureParams holds the input for passive memory capture.
-type PassiveCaptureParams struct {
-	SessionID string `json:"session_id"`
-	Content   string `json:"content"`
-	Project   string `json:"project,omitempty"`
-	Source    string `json:"source,omitempty"` // e.g. "subagent-stop", "session-end"
-}
+type PassiveCaptureParams = types.PassiveCaptureParams
 
 // PassiveCaptureResult holds the output of passive memory capture.
-type PassiveCaptureResult struct {
-	Extracted  int `json:"extracted"`  // Total learnings found in text
-	Saved      int `json:"saved"`      // New observations created
-	Duplicates int `json:"duplicates"` // Skipped because already existed
-}
+type PassiveCaptureResult = types.PassiveCaptureResult
 
 // learningHeaderPattern matches section headers for learnings in both English and Spanish.
 var learningHeaderPattern = regexp.MustCompile(
